refactor(cli): Split main.rs (1365 LOC) into cli_args, app_runtime, print_mode modules
Decompose the monolithic main.rs into focused modules using Rust's split
impl blocks:

- cli_args.rs (173 LOC): Cli struct with 27 clap fields and Commands enum
- app_runtime.rs (495 LOC): Runtime lifecycle (run, subcommands, hooks,
  interactive, stdin, session saving)
- print_mode.rs (296 LOC): One-shot print mode with API calls, fallback
  models, and output formatting
- main.rs (438 LOC): App struct definition, App::new() with helpers
  (init_logging, load_settings, init_hooks, load_plugins, resolve_session),
  PluginState struct, and main() entry point

All App fields changed to pub(crate) to enable cross-file impl blocks.
All 84 TUI regression tests pass. Clippy clean.
